+++ src/core/util/tag_state.rs
@@ -4,6 +4,8 @@ use std::collections::HashMap;
 use std::error::Error;
 use std::fmt::{Display, Formatter};
 
+/// Used to track `tag` directives
+#[derive(Debug)]
 pub struct TagState {
     listening: Option<String>,
     stored: HashMap<String, String>,
@@ -31,18 +33,18 @@ impl TagState {
     pub fn create(&mut self, tag: &str) -> Result<(), TagStateError> {
         if let Some(old_tag) = &self.listening {
             return Err(Report::new(TagStateError).attach_printable(format!(
-                "Cannot create new tag {tag} when old tag {old_tag} is still listening"
+                "Cannot create new tag `{tag}` when old tag `{old_tag}` is still listening."
             )));
         }
         for k in self.stored.keys() {
             if k.starts_with(tag) || tag.starts_with(k) {
                 if k == tag {
                     return Err(Report::new(TagStateError).attach_printable(format!(
-                        "Cannot create new tag {tag} when a tag with the same name is still listening"
+                        "Cannot create new tag `{tag}` because a tag exists with the same name."
                     )));
                 }
                 return Err(Report::new(TagStateError).attach_printable(format!(
-                    "Cannot create new tag {tag} because it is ambiguous with existing tag {k}"
+                    "Cannot create new tag `{tag}` because it is ambiguous with existing tag `{k}`."
                 )));
             }
         }
@@ -63,6 +65,8 @@ impl TagState {
     }
 
     pub fn inject_tags(&mut self, output: &str, line_ending: &str) -> String {
+        // argument should never have newline
+        assert!(!output.ends_with('\n'));
         let mut to_inject = self
             .stored
             .iter()
@@ -86,26 +90,43 @@ impl TagState {
             self.stored.remove(&key);
         }
         injected_output.push_str(&output[last_end..]);
-        injected_output.push_str(line_ending);
         injected_output
+    }
+
+    pub fn has_tags(&self) -> bool {
+        self.listening.is_some() || !self.stored.is_empty()
     }
 }
 
+impl Display for TagState {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        let mut tags = vec![];
+        if let Some(listening) = &self.listening {
+            tags.push(listening.clone());
+        }
+        tags.extend(self.stored.keys().cloned());
+        write!(f, "{}", tags.join(" ,"))
+    }
+}
 #[cfg(test)]
 mod tests {
     use super::*;
     #[test]
     fn test_create_ok() {
         let mut tag_state = TagState::new();
+        assert!(!tag_state.has_tags());
         tag_state.create("tag1").unwrap();
         assert_eq!(Some("tag1".to_string()), tag_state.listening);
+        assert!(tag_state.has_tags());
     }
 
     #[test]
     fn test_create_already_listening() {
         let mut tag_state = TagState::new();
+        assert!(!tag_state.has_tags());
         tag_state.create("tag1").unwrap();
         assert!(tag_state.create("tag2").is_err());
+        assert!(tag_state.has_tags());
     }
 
     #[test]
@@ -156,6 +177,7 @@ mod tests {
             Some("content1".to_string()),
             tag_state.stored.get("tag1").cloned()
         );
+        assert!(tag_state.has_tags());
     }
 
     #[test]
@@ -179,15 +201,16 @@ mod tests {
     fn test_tag_inject_one_all() {
         let mut tag_state = create_test_state();
         let output = tag_state.inject_tags("tag1", "\n");
-        assert_eq!("content1\n", &output);
+        assert_eq!("content1", &output);
         assert_eq!(2, tag_state.stored.len());
+        assert!(tag_state.has_tags());
     }
 
     #[test]
     fn test_tag_inject_one_prefixed() {
         let mut tag_state = create_test_state();
         let output = tag_state.inject_tags("hellotag1", "\n");
-        assert_eq!("hellocontent1\n", &output);
+        assert_eq!("hellocontent1", &output);
         assert_eq!(2, tag_state.stored.len());
     }
 
@@ -195,7 +218,7 @@ mod tests {
     fn test_tag_inject_one_postfixed() {
         let mut tag_state = create_test_state();
         let output = tag_state.inject_tags("tag1hello", "\n");
-        assert_eq!("content1hello\n", &output);
+        assert_eq!("content1hello", &output);
         assert_eq!(2, tag_state.stored.len());
     }
 
@@ -203,7 +226,7 @@ mod tests {
     fn test_tag_inject_one_sandwiched() {
         let mut tag_state = create_test_state();
         let output = tag_state.inject_tags("worldtag1hello", "\n");
-        assert_eq!("worldcontent1hello\n", &output);
+        assert_eq!("worldcontent1hello", &output);
         assert_eq!(2, tag_state.stored.len());
     }
 
@@ -211,7 +234,7 @@ mod tests {
     fn test_tag_inject_one_sandwiched_newlines() {
         let mut tag_state = create_test_state();
         let output = tag_state.inject_tags("worldtag1hello", "\n");
-        assert_eq!("worldcontent1hello\n", &output);
+        assert_eq!("worldcontent1hello", &output);
         assert_eq!(2, tag_state.stored.len());
     }
 
@@ -219,7 +242,7 @@ mod tests {
     fn test_tag_inject_many() {
         let mut tag_state = create_test_state();
         let output = tag_state.inject_tags("tag1tag2tag3", "\n");
-        assert_eq!("content1content2content3\n", &output);
+        assert_eq!("content1content2content3", &output);
         assert_eq!(0, tag_state.stored.len());
     }
 
@@ -227,7 +250,7 @@ mod tests {
     fn test_tag_inject_none() {
         let mut tag_state = create_test_state();
         let output = tag_state.inject_tags("asdftag 1fg   tag 2tag 3asdf", "\n");
-        assert_eq!("asdftag 1fg   tag 2tag 3asdf\n", &output);
+        assert_eq!("asdftag 1fg   tag 2tag 3asdf", &output);
         assert_eq!(3, tag_state.stored.len());
     }
 
@@ -235,7 +258,7 @@ mod tests {
     fn test_tag_inject_many_complex() {
         let mut tag_state = create_test_state();
         let output = tag_state.inject_tags("asdftag 1fg   tag2tag3asdf", "\n");
-        assert_eq!("asdftag 1fg   content2content3asdf\n", &output);
+        assert_eq!("asdftag 1fg   content2content3asdf", &output);
         assert_eq!(1, tag_state.stored.len());
     }
 
@@ -250,7 +273,7 @@ mod tests {
         tag_state.try_store("content3\n\n").unwrap();
         let output = tag_state.inject_tags("asdftag1fg   tag2tag3asdf", "\r\n");
         assert_eq!(
-            "asdfcont\r\nent1fg   \r\ncontent2\r\ncontent3\r\n\r\nasdf\r\n",
+            "asdfcont\r\nent1fg   \r\ncontent2\r\ncontent3\r\n\r\nasdf",
             &output
         );
         assert_eq!(0, tag_state.stored.len());