+++ tests/test.rs
@@ -10,25 +10,25 @@ testit!(examples__empty_test, |env| {
 });
 
 testit!(examples__include, |env| {
-    env.cfg().mode = Mode::Verify;
+    env.cfg.mode = Mode::Verify;
     // before output
     assert!(env.run().is_err());
-    env.cfg().mode = Mode::Build;
+    env.cfg.mode = Mode::Build;
     assert!(env.run().is_ok());
     env.assert_file_eq("foo.txt", "foo.txt.expected");
-    env.assert_file_eq("biz.txt", "bar.txt");
-    env.cfg().mode = Mode::Verify;
+    env.assert_file_eq("biz.txt", "biz.txt.expected");
+    env.cfg.mode = Mode::Verify;
     assert!(env.run().is_ok());
     env.set_file("foo.txt", "not the same");
     assert!(env.run().is_err());
-    env.cfg().mode = Mode::Clean;
+    env.cfg.mode = Mode::Clean;
     assert!(env.run().is_ok());
     env.assert_path_exists("foo.txt", false);
     // non-existent target should error
-    env.cfg().mode = Mode::Build;
-    env.cfg().inputs = vec!["a.txt".to_string()];
+    env.cfg.mode = Mode::Build;
+    env.cfg.inputs = vec!["a.txt".to_string()];
     assert!(env.run().is_err());
-    env.cfg().inputs = vec!["a.txtpp".to_string()];
+    env.cfg.inputs = vec!["a.txtpp".to_string()];
     assert!(env.run().is_err());
 });
 
@@ -49,17 +49,17 @@ testit!(examples__multiple_include_windows, |env| {
 });
 
 testit!(examples__run, |env| {
-    env.cfg().mode = Mode::Build;
-    env.cfg().inputs = vec!["foo.txt.txtpp".to_string()];
+    env.cfg.mode = Mode::Build;
+    env.cfg.inputs = vec!["foo.txt.txtpp".to_string()];
     assert!(env.run().is_ok());
     env.assert_file_eq("foo.txt", "foo.txt.expected");
-    env.cfg().inputs = vec!["invalid.txtpp".to_string()];
+    env.cfg.inputs = vec!["invalid.txtpp".to_string()];
     assert!(env.run().is_err());
 });
 
 testit!(examples__circular_dep, |env| {
-    env.cfg().mode = Mode::Build;
-    env.cfg().inputs = vec!["a.txt".to_string()];
+    env.cfg.mode = Mode::Build;
+    env.cfg.inputs = vec!["a.txt".to_string()];
     assert!(env.run().is_err());
     env.assert_file_eq("c.txt", "c.txt.expected");
 });
@@ -70,18 +70,16 @@ testit!(examples__circular_dep_self, |env| {
 });
 
 testit!(examples__temp__write_clean, |env| {
-    env_logger::init();
-
     assert!(env.run().is_ok());
     // content should match
     env.assert_file_eq("exp", "exp.expected");
     env.assert_file_eq("temp.out", "temp.out.expected");
-    env.cfg().mode = Mode::Clean;
+    env.cfg.mode = Mode::Clean;
     assert!(env.run().is_ok());
     env.assert_path_exists("temp.out", false);
     env.assert_path_exists("exp", false);
     // verify should fail fast, so temp.out is not created
-    env.cfg().mode = Mode::Verify;
+    env.cfg.mode = Mode::Verify;
     assert!(env.run().is_err());
     env.set_file("exp", "not the same");
     assert!(env.run().is_err());
@@ -99,10 +97,18 @@ testit!(examples__temp__python_script, |env| {
     assert!(env.run().is_ok());
     env.assert_file_eq("city.js", "city.js.expected");
     env.assert_file_eq("gen_cities.g.py", "gen_cities.g.py.expected");
-    env.cfg().mode = Mode::Clean;
+    env.cfg.mode = Mode::Clean;
     assert!(env.run().is_ok());
     env.assert_path_exists("city.js", false);
     env.assert_path_exists("gen_cities.g.py", false);
+});
+
+testit!(examples__trailing_newline, |env| {
+    assert!(env.run().is_ok());
+    env.assert_file_eq("source", "source.expected.trailing");
+    env.cfg.trailing_newline = false;
+    assert!(env.run().is_ok());
+    env.assert_file_eq("source", "source.expected");
 });
 
 testit!(examples__tag__inject_one, |env| {
@@ -138,4 +144,27 @@ testit!(examples__tag__no_output_skip, |env| {
     assert!(env.run().is_ok());
     env.assert_file_eq("test", "test.expected");
     env.assert_file_eq("temp", "temp.expected");
+});
+
+testit!(examples__tag__error, |env| {
+    env.cfg.inputs = vec!["tag_multiple_same".to_string()];
+    assert!(env.run().is_err());
+    env.cfg.inputs = vec!["tag_multiple_prefix".to_string()];
+    assert!(env.run().is_err());
+    env.cfg.inputs = vec!["tag_multiple_prefix2".to_string()];
+    assert!(env.run().is_err());
+});
+
+testit!(examples__tag__inject_once, |env| {
+    assert!(env.run().is_ok());
+    env.assert_file_eq("test", "test.expected");
+});
+
+testit!(examples__tag__unused_at_end, |env| {
+    assert!(env.run().is_err());
+});
+
+testit!(examples__tag__complex, |env| {
+    assert!(env.run().is_ok());
+    env.assert_file_eq("test", "test.expected");
 });